backend/requirements.txt
asgiref==3.11.1
Django==6.0.4
djangorestframework==3.17.1
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.4.0
